Switched to wxPython3 and sqlalchemy0.9. Reworking the code to fit the new packages.

+++ lib/psysmon/core/packageSystem.py
@@ -637,7 +637,7 @@ class Package:
                                   {'mysql_engine': 'InnoDB'}
                                  )
 
-                id = Column(Integer(10), primary_key=True, autoincrement=True)
+                id = Column(Integer, primary_key=True, autoincrement=True)
                 serial = Column(String(45), nullable=False)
                 type = Column(String(255), nullable=False)
                 description = Column(String(255), nullable=True)
